### 2018-04-16 #677 @whedon
[PRE REVIEW]: KLRfome - Kernel Logistic Regression on Focal Mean Embeddings
**Submitting author:** @mrecos (<a href="http://orcid.org/0000-0002-4627-7692">Matthew Harris</a>)
**Repository:** <a href="https://github.com/mrecos/klrfome" target ="_blank">https://github.com/mrecos/klrfome</a>
**Version:** v2.1.0
