### 2018-04-16 #677 @mrecos
Hello. Thanks for the Pre-review! From the reviewer list I would suggest: @benmarwick , @jsta , and @nuest  or others with experience in archaeology, machine learning, or R and spatial analysis.
PS - I can certainly add more details to the paper.md if desired.
Thank you all for your efforts.
