### 2018-04-16 51994df
Included the SAA 2018 research poster in the README
