### 2018-04-19 92a32fd
fixed bug in make_quads so it actually returns the quantities
