### 2018-04-19 f36000b
script updates
