### 2018-04-23 #677 @arfon
> PS - I can certainly add more details to the paper.md if desired.

Thanks @mrecos - yes, please try and flesh the paper out in a little more detail following the guidelines from the main JOSS site:

> - A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
> - A clear statement of need that illustrates the purpose of the software
> - A list of key references including a link to the software archive
> - Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
