### 2018-04-23 #677 @arfon
@benmarwick - would you be willing to review this submission for JOSS?
