### 2018-04-23 de5cf7a
README update with thank yous and references section linked to ./JOSS/paper.bib
