+++ README.md
@@ -12,12 +12,14 @@ The purpose of this package is to solve the *Distribution Regression* problem fo
 
 To achieve this goal, the package fits a Kernel Logistic Regression (KLR) model onto a mean embedding similarity matrix and predicts as a roving focal function of varying window size. The name of the package is derived from this approach; **K**ernel **L**ogistic **R**egression on **FO**cal **M**ean **E**mbeddings (**klrfome**) pronounced *clear foam*.
 
+This model is inspired by and borrows from Zoltán Szabó's work on mean embeddings (Szabó et al. 2016, Szabó et al. (2015)) and Ji Zhu & Trevor Hastie's Kernel Logistic Regression algorithm (Zhu and Hastie 2005). I extend a hardy **thank you** to Zoltán for his correspondence during the development of this approach. However, all errors, oversights, and omissions are my own.
+
 <p align="center">
 <img width="800" height="600" src="https://github.com/mrecos/klrfome/blob/master/SAA_2018_poster/SAA_2018_poster_small.jpg?raw=true">
 </p>
 
-(High-res versions of research poster are in the [/SAA\_2018\_poster](https://github.com/mrecos/klrfome/tree/master/SAA_2018_poster) folder)
 
+(High-res versions of research poster are in the [/SAA\_2018\_poster](https://github.com/mrecos/klrfome/tree/master/SAA_2018_poster) folder)
 
 ### Citation
 
@@ -296,7 +298,7 @@ This package contains the functions necessary to compute Kernel Linear Regressio
 
 #### raster\_predict\_funs.R for Predicting on Raster Data
 
--   `rescale_sim_raster` - Function that rescales simulated rasters from `NLMR::nlm_gaussianfield` or whatever you want to use, to the mean and standard deviation of the simulated data used to fit the `klr` model. You will have to add the `mean` and `sd` arguments manually based on what you put into the `get_sim_data` function. The example in the code above inputs the default `mean` and `sd` values from the defualts of the `get_sim_data` function. Returned is a raster scaled too your simualted training data.
+-   `rescale_sim_raster` - Function that rescales simulated rasters from `NLMR::nlm_gaussianfield` (Sciaini, Fritsch, and Simpkins 2017) or whatever you want to use, to the mean and standard deviation of the simulated data used to fit the `klr` model. You will have to add the `mean` and `sd` arguments manually based on what you put into the `get_sim_data` function. The example in the code above inputs the default `mean` and `sd` values from the defualts of the `get_sim_data` function. Returned is a raster scaled too your simualted training data.
 
[…]
