### 2018-04-23 a9d1b16
JOSS paper update - added more details and package information to text.
