+++ JOSS/paper.md
@@ -21,10 +21,19 @@ affiliations:
 
 # Summary
 
-- KLRfome is a package for predicting a geographic area's realtive sensitivity for the presence of archaeological sites. This is achieved by fitting, predicting, and visualizing a Kernel Logistic Regression model on mean feature embeddings. This regression algorithm and package are created to improve upon the current methods in archaeological predictive modeling. These improvements include 1) modeling rich descriptions of archaeological landforms by mitigating undesiarable spatial corraltion between samples, 2) explicitly modeling the similarity between archaeological sites as characterized by rich features, 3) the ability to define research specific similairty measures, and 4) focal window prediction that can be modified based on theory or managment goals. This work is based on Szabó's work on feature mean embeddings for distribution regression [@Szabo] and the Kernel Logistic Regression algortihm in [@Zhu]. The KLRfome package is available at [https://github.com/mrecos/klrfome](https://github.com/mrecos/klrfome) and DOI [https://doi.org/10.5281/zenodo.1218403](https://doi.org/10.5281/zenodo.1218403)
+KLRfome is a package for predicting a geographic area's realtive sensitivity for the presence of archaeological sites. This is achieved by fitting, predicting, and visualizing a Kernel Logistic Regression model on mean feature embeddings. This regression algorithm and package are created to improve upon the current methods in archaeological predictive modeling. These improvements include 1) modeling rich descriptions of archaeological landforms by mitigating undesiarable spatial corraltion between samples, 2) explicitly modeling the similarity between archaeological sites as characterized by rich features, 3) the ability to define research specific similairty measures, and 4) focal window prediction that can be modified based on theory or managment goals. 
+
+More specifically, this package seeks to solve the *Distribution Regression* problem for archaeological site location modeling; or any other data for that matter. The aim of Distribution Regression is to map a single scalar outcome (e.g. presence/absence; 0/1) to a distribution of features. This is opposed to typical regression where you have one observation mapping a single outcome to a single set of features/predictors. For example, an archaeological site is singularly defined as either present or absent, however the area within the sites boundary is not singularly defined by any one measurement. The area with an archaeology site is defined by an infinite distribution of measurements. Modeling this in traditional terms means either collapsing that distribution to a single measurement or pretending that a site is actually a series of adjacent, but independent measurements. The methods developed for this package take a different view instead by modeling the distribution of measurements from within a single site on a scale of similarity to the distribution of measurements on other sites and the environmental background in general. This method avoids collapsing measurements and promotes the assumption of independence from within a site to between sites. By doing so, this approach models a richer description of the landscape in a more intuitive sense of similarity.
+
+To achieve this goal, the package fits a Kernel Logistic Regression (KLR) model onto a mean embedding similarity matrix and predicts as a roving focal function of varying window size. The name of the package is derived from this approach; **K**ernel **L**ogistic **R**egression on **FO**cal **M**ean **E**mbeddings (**klrfome**) pronounced *clear foam*. This work is based on Szabó's work on feature mean embeddings for distribution regression [@Szabo, @Szabo2] and the Kernel Logistic Regression algortihm in [@Zhu]. A [poster](https://github.com/mrecos/klrfome/tree/master/SAA_2018_poster) introducing the KLRfome package was presented on April 12, 2018 at the Society for American Archaeology annual meeting in Washington, D.C.The KLRfome package is available at [https://github.com/mrecos/klrfome](https://github.com/mrecos/klrfome) and DOI [https://doi.org/10.5281/zenodo.1218403](https://doi.org/10.5281/zenodo.1218403)
+
+
[…]
