+++ JOSS/paper.bib
@@ -1,17 +1,83 @@
 @article{Szabo,
-   author={Szabó, Zoltán and Bharath Sriperumbudur and Barnabás Póczos and Arthur Gretton},
+   author={Szabó, Zoltán and Sriperumbudur, Bharath and Póczos, Barnabás and Gretton, Arthur},
    title={Learning theory for distribution regression},
    journal={Journal of Machine Learning Research},
    volume={17},
-   year=2016,
+   year={2016},
    pages={1-40}
 }
 
 @article{Zhu,
-   author={Zhu, Ji and Trevor Hastie},
+   author={Zhu, Ji and Hastie, Trevor},
    title={Kernel logistic regression and the import-vector machine},
    journal={Journal of Computational and Graphical Statistics},
-   volume={14}},
-   year=2005,
+   volume={14},
+   number={1},
+   year={2005},
    pages={185-205}
 }
+
+@article{Muandet,
+   author={Muandet, Krikamol and Fukumizu, Kenji and Sriperumbudur, Bharath and Schölkopf, Bernhard},
+   title={Kernel Mean Embedding of Distributions: A Review and Beyond},
+   journal={Foundations and Trends in Machine Learning},
+   volume={10},
+   number={1-2},
+   year={2017},
+   pages={1-141}
+}
+
+@incollection{Flaxman,
+   author={Flaxman, Seth and Yu-Xiang Wang and Smola, Alexander J.},
+   title={Who Supported Obama in 2012?: Ecological inference through distribution regression},
+   booktitle={Proceedings of the 21th ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+   publisher = {ACM},
+   year={2015},
+   pages={289–298}
+}
+
+@ARTICLE{2017arXiv170504293L,
+   author = {{Law}, H.~C.~L. and {Sutherland}, D.~J. and {Sejdinovic}, D. and 
+	{Flaxman}, S.},
+    title = "{Bayesian Approaches to Distribution Regression}",
+  journal = {ArXiv e-prints},
+archivePrefix = "arXiv",
+   eprint = {1705.04293},
+ primaryClass = "stat.ML",
+ keywords = {Statistics - Machine Learning, Computer Science - Learning},
+     year = {2017},
+    month = {may},
+   adsurl = {http://adsabs.harvard.edu/abs/2017arXiv170504293L},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@software{NLMR,
+  author = {Sciaini, Marco and Fritsch, Matthias and Simpkins, Craig E.},
+  title = {NLMR: Simulating neutral landscape models},
+  url = {https://CRAN.R-project.org/package=NLMR},
+  version = {0.2.0},
+  date = {2017-03-25},
+}
+
+@article{Song,
+   author={Song, Li and Fukumizu, Kenjiand Gretto, Arthur},
+   title={ Kernel Embeddings of Conditional Distributions: A Unified Kernel Framework for Nonparametric Inference in Graphical Models},
+   journal={IEEE Signal Processing Magazine},
+   volume={30},
+   number={4},
+   year={2013},
+   pages={98-111}
+}
+
+@incollection{Szabo2,
+   author={Szabó, Zoltán and Gretton, Arthur and Póczos, Barnabás and Sriperumbudur, Bharath},
+   title={Two-stage sampled learning theory on distributions},
+   booktitle={International Conference on Artificial Intelligence and Statistics (AISTATS)},
+   year={2015},
+   pages={948–957}
+}
+
+
+
+
+
