### 2018-04-23 #677 @mrecos
Thank you @arfon. I added a bit more detail to the paper. Please let me know if you have other suggested edits prior to review.
