### 2018-05-01 #677 @leeper
@arfon would you like me to handle as editor?
