### 2018-05-03 #722 @whedon
[REVIEW]: KLRfome - Kernel Logistic Regression on Focal Mean Embeddings
**Submitting author:** @mrecos (<a href="http://orcid.org/0000-0002-4627-7692">Matthew Harris</a>)
**Repository:** <a href="https://github.com/mrecos/klrfome" target ="_blank">https://github.com/mrecos/klrfome</a>
**Version:** v2.2.0

BM: I think Szabo should be noted in the readme as someone who made a contribution that made this pkg possible
BM: done

BM: very nicely stated.
BM: There is no vignette, but the README gives a very thorough walk-through. I think I've seen a method to have a vignette, and auto-extract it into the README on pkg build, that would help to fulfil this requirement.
BM: done
BM: documentation is present, but uneven. Descriptions and examples are missing for many functions. These should be added. Spelling: scaler should be scalar. The KLR function has a bit of commented-out code in it that probably should be removed.
BM: done
BM: not yet, though the readme is a good start.
BM: done
BM: There is no 'Maintainer' field in description. The pkg referred to in the CONTRIBUTING is DistRegLMERR, presumably an earlier name for this pkg, but should be changed.
BM: done

### Additional criteria (from rOpenSci)
BM: I pasted these over from the rOpenSci onboarding
 - [x] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines: https://github.com/ropensci/onboarding/blob/master/packaging_guide.md
 BM: not yet, still some documentation to add. There is no NEWS file. There are no tests.
BM: done
BM: not yet.
BM: done

   3
   ---

The author has done an excellent job of fulfilling the requirements for JOSS. In particular, it's an excellent example of superb package documentation to the point that I believe any undergraduate archaeology student (or related field) should be able to make sense of what is the need and purpose, and quickly start using the pkg. The vignette is outstanding and deserves a wide audience. Well done!
