### 2018-05-18 #722 @leeper
@benmarwick I can see you've made some progress on this review. Can you open issues over at [the source repo](https://github.com/mrecos/klrfome/issues) for any issues you'd like to see addressed? Thanks!
