### 2018-05-18 mrecos/klrfome#1 @benmarwick
More documentation needed
It would be good to see some more detail in the help pages, and a vignette.

See my comments here: https://github.com/openjournals/joss-reviews/issues/722#issue-320042175
