### 2018-05-18 mrecos/klrfome#2 @benmarwick
More examples needed
Each exported fn needs an example in its help pages. See my comments here: https://github.com/openjournals/joss-reviews/issues/722#issue-320042175
