### 2018-05-18 mrecos/klrfome#3 @benmarwick
Tests needed
More details: https://github.com/openjournals/joss-reviews/issues/722#issue-320042175
