### 2018-05-18 mrecos/klrfome#4 @benmarwick
NEWS needed 
cf. https://github.com/openjournals/joss-reviews/issues/722#issue-320042175
