### 2018-05-21 #722 @leeper
@benmarwick Great. Thanks! Let us know if you have any other comments at this point.

@mrecos Please try to address the issues opened by @benmarwick. The matters of examples and tests are particularly important.
