### 2018-05-22 #722 @mrecos
Great! Thanks @benmarwick and @leeper. I will address in short order and report back.
