### 2018-08-08 06587c5
Readme update with special Thanks. contributes to Issue #1
