+++ README.md
@@ -1,31 +1,72 @@
 
-[![DOI](https://zenodo.org/badge/103055953.svg)](https://zenodo.org/badge/latestdoi/103055953) [![Build Status](https://travis-ci.org/mrecos/klrfome.svg?branch=master)](https://travis-ci.org/mrecos/klrfome)
+[![DOI](https://zenodo.org/badge/103055953.svg)](https://zenodo.org/badge/latestdoi/103055953)
+[![Build
+Status](https://travis-ci.org/mrecos/klrfome.svg?branch=master)](https://travis-ci.org/mrecos/klrfome)
 
 <p align="center">
-<img width="326" height="134" src="https://github.com/mrecos/klrfome/blob/master/klrfome_logo/KLR-black.png?raw=true">
-</p>
 
+<img width="326" height="134" src="https://github.com/mrecos/klrfome/blob/master/klrfome_logo/KLR-black.png?raw=true">
+
+</p>
 
 #### klrfome - Kernel Logistic Regression on Focal Mean Embeddings
 
-The purpose of this package is to solve the *Distribution Regression* problem for archaeological site location modeling; or any other data for that matter. The aim of Distribution Regression is to map a single scalar outcome (e.g. presence/absence; 0/1) to a distribution of features. This is opposed to typical regression where you have one observation mapping a single outcome to a single set of features/predictors. For example, an archaeological site is singularly defined as either present or absent, however the area within the sites boundary is not singularly defined by any one measurement. The area with an archaeology site is defined by an infinite distribution of measurements. Modeling this in traditional terms means either collapsing that distribution to a single measurement or pretending that a site is actually a series of adjacent, but independent measurements. The methods developed for this package take a different view instead by modeling the distribution of measurements from within a single site on a scale of similarity to the distribution of measurements on other sites and the environmental background in general. This method avoids collapsing measurements and promotes the assumption of independence from within a site to between sites. By doing so, this approach models a richer description of the landscape in a more intuitive sense of similarity.
+The purpose of this package is to solve the *Distribution Regression*
+problem for archaeological site location modeling; or any other data for
+that matter. The aim of Distribution Regression is to map a single
+scalar outcome (e.g. presence/absence; 0/1) to a distribution of
+features. This is opposed to typical regression where you have one
+observation mapping a single outcome to a single set of
+features/predictors. For example, an archaeological site is singularly
+defined as either present or absent, however the area within the sites
+boundary is not singularly defined by any one measurement. The area with
+an archaeology site is defined by an infinite distribution of
+measurements. Modeling this in traditional terms means either collapsing
+that distribution to a single measurement or pretending that a site is
+actually a series of adjacent, but independent measurements. The methods
+developed for this package take a different view instead by modeling the
+distribution of measurements from within a single site on a scale of
+similarity to the distribution of measurements on other sites and the
+environmental background in general. This method avoids collapsing
+measurements and promotes the assumption of independence from within a
+site to between sites. By doing so, this approach models a richer
+description of the landscape in a more intuitive sense of similarity.
 
-To achieve this goal, the package fits a Kernel Logistic Regression (KLR) model onto a mean embedding similarity matrix and predicts as a roving focal function of varying window size. The name of the package is derived from this approach; **K**ernel **L**ogistic **R**egression on **FO**cal **M**ean **E**mbeddings (**klrfome**) pronounced *clear foam*.
-
[…]
