### 2018-08-08 #722 @mrecos
Thanks @leeper - Making edits/additions/commits now.
