### 2018-08-08 f4c3f3f
Added NEWS file for package updates. Applies to Issue #4
