### 2018-08-08 mrecos/distreglmerr#4 @mrecos
Added NEWS file
