### 2018-08-08 814e846
documentation updates aligned with Issue #1
