+++ DESCRIPTION
@@ -21,4 +21,4 @@ Imports:
 Suggests:
   boot,
   pROC,
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
