### 2018-08-08 7079348
Documentation update - get_sim_data.R. Aligned with Issue #1
