+++ R/format_site_data.R
@@ -28,7 +28,7 @@
 #' train_presence <- formatted_data[["train_presence"]]
 #' test_presence <- formatted_data[["test_presence"]]
 #'
-#' ##### Logistic Mean Embedding KRR Model
+#' ##### Logistic Mean Embedding KLR Model
 #' #### Build Kernel Matrix
 #' K <- build_K(train_data, sigma = sigma, dist_metric = dist_metric)
 #' #### Train
