+++ NAMESPACE
@@ -10,7 +10,6 @@ export(build_K)
 export(crop_raster_collar)
 export(format_site_data)
 export(get_k)
-export(get_sim_data)
 export(make_quads)
 export(metrics)
 export(rescale_sim_raster)
