### 2018-08-08 2bc7b4c
Documentation update for KLR functions. Aligned with Issue #1
