+++ R/KLR_funs.R
@@ -1,4 +1,8 @@
 #' KLR
+#' 
+#' `KLR()` Is the primary model fitting function in the `klrfome` package. This function fits a Kernel Logistic Regression (KLR) model to a similarity or distance matrix.
+#' 
+#' the `KLR` function takes the similarity kernel matrix `K`, a vector `presnece` of presence/absence coded as 1 or 0, and a scalar value for the `lambda` regularizing hyperparameter; optionally values for maximum iterations and threshold. This function performs Kernel Logistic Regression (KLR) via Iterative Re-weighted Least Squares (IRLS). The objective is to approximate a set of parameters that minimize the negative likelihood of the parameters given the data and response. This function returns a list of `pred`, the estimated response (probability of site-presence) for the training data, and `alphas`, the approximated parameters from the IRLS algorithm.
 #'
 #' @param K - [NxN] Mean embedding kernel matrix
 #' @param presence - [vector] or presence = 1 or absence = 0
@@ -10,11 +14,32 @@
 #' @return list: `pred` - predicted probabiity of positive class, `alphas` - estimated coefficients
 #' @importFrom Matrix Diagonal
 #' @export
+#' @examples
+#'\dontrun{
+#' sim_data <- get_sim_data(site_samples = 800, N_site_bags = 75,
+#' sites_var1_mean = 80, sites_var1_sd   = 10,
+#' sites_var2_mean = 5,  sites_var2_sd   = 2,
+#' backg_var1_mean = 100,backg_var1_sd   = 20,
+#' backg_var2_mean = 6,  backg_var2_sd   = 3)
+#' formatted_data <- format_site_data(sim_data, N_sites=10, train_test_split=0.8,
+#'                                    sample_fraction = 0.9, background_site_balance=1)
+#' train_data <- formatted_data[["train_data"]]
+#' train_presence <- formatted_data[["train_presence"]]
+#' test_presence <- formatted_data[["test_presence"]]
+#'
+#' ##### Logistic Mean Embedding KLR Model
+#' #### Build Kernel Matrix
+#' K <- build_K(train_data, sigma = sigma, dist_metric = dist_metric)
+#' #### Train
+#' train_log_pred <- KLR(K, train_presence, lambda, 100, 0.001, verbose = 2)
+#' #### Predict
+#' test_log_pred <- KLR_predict(test_data, train_data, dist_metric = dist_metric,
+#'                             train_log_pred[["alphas"]], sigma)
+#'}
 #'
 KLR <- function(K, presence, lambda, maxiter = 100, tol = 0.01, verbose=1){
-  # LOGISTIC - optimize alpha
   if(is.vector(K)){
-    N = length(K)# NOT SURE IF THIS WORKS WITH REST OF FUNCTION
+    N = length(K)
   } else if(is.matrix(K)){
     N = nrow(K)
   }
@@ -25,7 +50,6 @@ KLR <- function(K, presence, lambda, maxiter = 100, tol = 0.01, verbose=1){
     spec = 1 + exp(-Kalpha)
     pi = 1 / spec
     diagW = pi * (1 - pi)
-    # e = (presence - pi) / diagW
     # same as: Zhu & Hastie (2002:19)
     # W^-1(y-pi); W = diag[pi(1 − pi)]N×N
     # e = solve(as.matrix(Matrix::Diagonal(x=diagW))) %*% (presence - pi)
@@ -41,7 +65,6 @@ KLR <- function(K, presence, lambda, maxiter = 100, tol = 0.01, verbose=1){
     alphan = as.vector(alpha_new)
     if(verbose == 2){
       cat("Step ", iter, ". Absolute Relative Approximate Error = ",
-          # round(sum(abs((alphan - alpha)/alphan))*100, 4), "\n", sep = "")
[…]
