### 2019-01-16 b281fbd
further documeted metrics.R, deleted CI_metrics function, ran roxygen documentation, rebuilt package. Contributing to Issue #1
