+++ NAMESPACE
@@ -8,7 +8,6 @@ export(K_corrplot)
 export(build_K)
 export(crop_raster_collar)
 export(format_site_data)
-export(make_quads)
 export(metrics)
 export(rescale_sim_raster)
 export(scale_prediction_rasters)
