### 2019-01-16 708324d
Metrics.R documentation, Issue #1
