### 2019-01-16 a32ebf6
Documented plot.R function. Fixed roxygen typo for @examples. ran document & build
