+++ NAMESPACE
@@ -2,12 +2,15 @@
 
 export(CM_quads)
 export(KLR)
+export(KLR_predict)
 export(KLR_predict_each)
 export(KLR_raster_predict)
 export(K_corrplot)
 export(build_K)
 export(crop_raster_collar)
 export(format_site_data)
+export(get_sim_data)
+export(make_quads)
 export(metrics)
 export(rescale_sim_raster)
 export(scale_prediction_rasters)
