### 2019-01-16 458d0cd
Fixed package name in CONTRIBUTING.md related to Issue #1
