### 2019-01-16 6acf7f6
cleaned & added documentation. BUG FIX - in `pred_cat = ifelse(dat$pred >= threshold[i],1,0)` where `dat$pred` was `model_pred$pred`
