### 2019-01-16 4982071
added documentation to raster_predict_funs.R
