### 2019-01-16 6b1bd5d
corrected lines in `KLR_raster_predict()` to use `:::` for `KLR_predict_each()` so that an exported function can call a non-exported function. `klrfome:::KLR_predict_each()`. Documented and rebuilt.
