### 2019-01-16 mrecos/klrfome#1 @mrecos
All features are documented with titles, descriptions, parameters, notes, and examples.
