### 2019-01-16 mrecos/distreglmerr#1 @mrecos
All features are documented with titles, descriptions, parameters, notes, and examples. Closing this issue
