### 2019-01-16 mrecos/distreglmerr#2 @mrecos
Each exported function has an example. Most are tagged as `\dontrun` but do have fully functional examples. The issue is that to compute the data required as inputs to some function (prediction for example) require a good bit of computing. Closing issue.
