### 2019-01-16 #722 @mrecos
Hello @afron. Quick update: I took care of 2 of 3 issues on my repo. TODO - convert existing how-to to vignette and unit tests.
