### 2019-01-16 80f7c21
STARTED UNIT TESTING. ran `devtools::use_testthat()` to set up DESCRIPTION and ./tests folder structure. Wrote first tese to check output of `get_sim_data()`
