### 2019-02-08 #722 @mrecos
Hello @arfon . Going well. I took care of all points except for the unit tests. I am currently implementing those.
