### 2019-02-08 cb3b8d5
- added vignette and accompanying files/folders - edit to format_site_data.r for error checking - edit to KLR_raster_predict() in raster_predict_fun.R to reference non-exported split_raster_stack function
