+++ inst/doc/vignette_intro_and_example_workflow.R
@@ -1,0 +1,163 @@
+## ----setup, include=FALSE------------------------------------------------
+knitr::opts_chunk$set(echo = TRUE)
+
+## ---- echo=FALSE, warning=FALSE, out.width="800px"-----------------------
+knitr::include_graphics("https://github.com/mrecos/klrfome/blob/master/README_images/KLRfome_dataflow.png?raw=true")
+
+## ----packages, message=FALSE, warning=FALSE, paged.print=FALSE-----------
+library("ggplot2")   # for plotting results
+library("NLMR")      # for creating simulated landscapes
+library("rasterVis") # for plotting simulated lan
+library("pROC")      # for evaluation of model AUC metric
+library("dplyr")     # for data manipulation
+library("knitr")     # for printing tables in this document
+library("klrfome")   # for modeling
+
+## ----params--------------------------------------------------------------
+#Parameters
+set.seed(232)
+sigma = 0.5
+lambda = 0.1
+dist_metric = "euclidean"
+
+## ----sim_data------------------------------------------------------------
+### Simulate Training Data
+sim_data <- get_sim_data(site_samples = 800, N_site_bags = 75)
+formatted_data <- format_site_data(sim_data, N_sites=10, train_test_split=0.8,
+                                   sample_fraction = 0.9, background_site_balance=1)
+train_data <- formatted_data[["train_data"]]
+train_presence <- formatted_data[["train_presence"]]
+test_data <- formatted_data[["test_data"]]
+test_presence <- formatted_data[["test_presence"]]
+
+## ----fit_model-----------------------------------------------------------
+##### Logistic Mean Embedding KRR Model
+#### Build Kernel Matrix
+K <- build_K(train_data, sigma = sigma, dist_metric = dist_metric, progress = FALSE)
+#### Train KLR model
+train_log_pred <- KLR(K, train_presence, lambda, 100, 0.001, verbose = 2)
+#### Predict KLR model on test data
+test_log_pred <- KLR_predict(test_data, train_data, dist_metric = dist_metric,
+                             train_log_pred[["alphas"]], sigma, progress = FALSE)
+
+### Plot K Matrix
+K_corrplot(K,train_data,clusters=4)
+
+### Plot Test Set Prediction
+predicted_log <- data.frame(pred = test_log_pred, obs = test_presence)
+ggplot(predicted_log, aes(x = as.factor(obs), y = pred, color = as.factor(obs))) +
+  geom_jitter(width = 0.1) +
+  theme_bw() +
+  ylim(c(0,1)) +
+  labs(y = "Predicted Probability", x = "Site Presence",
+       title = "Kernel Logistic Regression",
+       subtitle = "test set predictions; simulated data") +
+  theme(
+    legend.position = "none"
+  )
+
+### Save parameters for later prediction
+params <- list(train_data = train_data,
+               alphas_pred = train_log_pred[["alphas"]],
+               sigma = sigma,
+               lambda = lambda,
+               means = formatted_data$means,
+               sds = formatted_data$sds)
+
+## ----echo=FALSE, warning=FALSE, out.width="800px"------------------------
+knitr::include_graphics("https://github.com/mrecos/klrfome/blob/master/README_images/KLRfome_prediction.png?raw=true")
+
+## ----predict_rasters-----------------------------------------------------
+### width and hieght of roving focal window (required)
+ngb = 5
+### Number of rows and columns in prediction rasters
+## needed for making simulated rasters, as well as for predicting real-world rasters
+cols = 100
[…]
