+++ inst/doc/vignette_intro_and_example_workflow.rmd
@@ -1,0 +1,295 @@
+---
+title: "klrfome - Introduction and example workflow"
+author: "Matthew Harris"
+date: "2/7/2019"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{klrfome - Introduction, example workflow, and function description}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+# klrfome - Kernel Logistic Regression on Focal Mean Embeddings
+
+## Introduction
+
+The purpose of this package is to solve the *Distribution Regression* problem for archaeological site location modeling; or any other data for that matter. The aim of Distribution Regression is to map a single scalar outcome (e.g. presence/absence; 0/1) to a distribution of features. This is opposed to typical regression where you have one observation mapping a single outcome to a single set of features/predictors. For example, an archaeological site is singularly defined as either present or absent, however the area within the sites boundary is not singularly defined by any one measurement. The area with an archaeology site is defined by an infinite distribution of measurements. Modeling this in traditional terms means either collapsing that distribution to a single measurement or pretending that a site is actually a series of adjacent, but independent measurements. The methods developed for this package take a different view instead by modeling the distribution of measurements from within a single site on a scale of similarity to the distribution of measurements on other sites and the environmental background in general. This method avoids collapsing measurements and promotes the assumption of independence from within a site to between sites. By doing so, this approach models a richer description of the landscape in a more intuitive sense of similarity.
+
+To achieve this goal, the package fits a Kernel Logistic Regression (KLR) model onto a mean embedding similarity matrix and predicts as a roving focal function of varying window size. The name of the package is derived from this approach; **K**ernel **L**ogistic **R**egression on **FO**cal **M**ean **E**mbeddings (**klrfome**) pronounced *clear foam*.
+
+
+## Example workflow on simulated data
+
+
+```{r, echo=FALSE, warning=FALSE, out.width="800px"}
+knitr::include_graphics("https://github.com/mrecos/klrfome/blob/master/README_images/KLRfome_dataflow.png?raw=true")
+```
+
+In brief, the process below is 1) take a table of observations of two or more environmental variables within known sites and across the background of the study area; 2) use `format_data()` to convert that table to a list and under-sample the background data to a desired ratio (each group of observations with a site or background area are referred o in the ML literature as "bags"); 3) use `build_k()` function with the `sigma` hyperparameter and distance (default `euclidean`) to create a similarity matrix between all site and background bags; 4) the similarity matrix is the object that the kernel logistic regression model `klr()` function uses to fit its parameters. Steps 3 and 4 are where this method detracts most from traditional regression, but it is also what sets this method apart. unlike most regression that fits a model to a table of measurements, this approach fits a model to a matrix of similarities between all of the units of analysis (sites and background areas).
+
+#### Libraries
+```{r packages, message=FALSE, warning=FALSE, paged.print=FALSE}
+library("ggplot2")   # for plotting results
+library("NLMR")      # for creating simulated landscapes
+library("rasterVis") # for plotting simulated lan
+library("pROC")      # for evaluation of model AUC metric
+library("dplyr")     # for data manipulation
+library("knitr")     # for printing tables in this document
+library("klrfome")   # for modeling
+```
+
+
+#### Set hyperparameters and load simulated site location data
[…]
