+++ DESCRIPTION
@@ -21,5 +21,8 @@ Imports:
 Suggests:
   boot,
   pROC,
-  testthat
+  testthat,
+  knitr, 
+  rmarkdown
 RoxygenNote: 6.1.0
+VignetteBuilder: knitr
