### 2019-02-18 4a3cd00
test_that unit tests for all major functions of the package. Applies to Issue #3
