### 2019-02-18 mrecos/distreglmerr#3 @mrecos
Completed unit testing with test_that for all major functions of the package and analytical process.
commit 4a3cd008dd40b20cbc662217d1cf2d88a29016a3
