### 2019-02-18 #722 @mrecos
Hello @arfon . I have completed the code updates requested by Ben via his review and Issues. These include general code cleanup, documentation of all public functions, a vignette, and unit testing. Please review and advise. Thanks!
