### 2019-02-18 #722 @arfon
> Hello @arfon . I have completed the code updates requested by Ben via his review and Issues. These include general code cleanup, documentation of all public functions, a vignette, and unit testing. Please review and advise. Thanks!

Great!

@benmarwick - when you get a chance, could you please come and take another look?
