### 2019-03-18 #722 @benmarwick
Thanks for the reminder. I've had another look at the package and everything looks great! The author has done an excellent job of fulfilling the requirements for JOSS. In particular, it's an excellent example of superb package documentation to the point that I believe any undergraduate archaeology student (or related field) should be able to make sense of what is the need and purpose, and quickly start using the pkg. The vignette is outstanding and deserves a wide audience. Well done!
