### 2019-03-18 #722 @labarba
I see some unchecked items in the checklist—if you're satisfied, can you tick off those items?
