### 2019-03-19 #722 @mrecos
Hello @arfon - I create a the "JOSS Release" and committed to Zenodo for a new DOI (10.5281/zenodo.2598673). I checked the DOI badge on the KLRfome repo and all seems well. Let me know if this covers it. Thanks!
