### 2019-03-19 #722 @labarba
Can you edit the metadata of the Zenodo deposit (no need to get new version or new DOI) so the title and author list match the JOSS paper?
