### 2019-03-19 5747977
update JOSS paper to fix typos
