### 2019-03-19 #722 @mrecos
Hello @labarba, ok updated! I had found a typo in the release version of the JOSS paper markdown just now. I pushed the edits to the repo, but they are now post 2.2.0 release. Should I do redo the release?
https://zenodo.org/record/2598673#.XJE_3FVKipo
