### 2019-03-19 #722 @labarba
If it's just an edit on the JOSS paper, I don't see a need to redo the release (the final JOSS paper is archived on its own in the journal).
