### 2019-03-19 #722 @labarba
Please check if the suggested "Mising DOIs" are indeed missing, or are flukes of `whedon`'s Crossref query.
