### 2019-03-19 cf2ccb9
updated JOSS bib with DOIs
