+++ JOSS/paper.bib
@@ -14,7 +14,8 @@
    volume={14},
    number={1},
    year={2005},
-   pages={185-205}
+   pages={185-205},
+   doi = {10.1198/106186005x25619}
 }
 
 @article{Muandet,
@@ -24,7 +25,8 @@
    volume={10},
    number={1-2},
    year={2017},
-   pages={1-141}
+   pages={1-141},
+   doi = {10.1561/2200000060}
 }
 
 @incollection{Flaxman,
@@ -66,7 +68,8 @@ archivePrefix = "arXiv",
    volume={30},
    number={4},
    year={2013},
-   pages={98-111}
+   pages={98-111},
+   doi = {10.1109/msp.2013.2252713}
 }
 
 @incollection{Szabo2,
