### 2019-03-19 #722 @mrecos
@labarba whedon was correct, those citations had DOIs to include. I commit/pushed a new paper.bib with the DOIs included. Aside from that, the proof looks good!
