### 2019-03-19 #722 @arfon
> The KLRfome package is available at https://github.com/mrecos/klrfome and DOI https://doi.org/10.5281/zenodo.1218403

@mrecos - could you please remove this sentence from your paper? We link to the repository and archive DOI in the paper (in the left-hand margin).
