### 2019-03-19 b18a403
Removed sentence from JOSS paper regarding DOI
