### 2019-03-19 #722 @mrecos
@arfon ok, sentence removed, committed, and pushed.
